+++ src/webpack/localVars.js
@@ -1,0 +1,8 @@
+export function addLocalVars(source, chunk, compilationHash) {
+  return [
+    source,
+    "// interleaving map",
+    "var interleaveMap = {};",
+    `var compilationHash = '${compilationHash}'`
+  ].join("\n");
+}
