+++ src/webpack/requireExtentions.js
@@ -1,13 +1,20 @@
-export const addInterleaveExtention = (source, chunk, webpackHash) => {
+export const addInterleaveExtention = (source, chunk, hash) => {
   return [
     source,
-    "// adding local registration function",
-    "function registerLocals(chunkMap) { console.log(webpackHash,chunkMap);//chunkMap[0].forEach(function(key){",
-    "//keys 0 map 1",
+    "// webpack chunk self registration",
+    `function registerLocals(chunkMap) { //chunkMap[0].forEach(function(key){`,
+    "var chunkCompileHash = chunkMap[0];",
+    "var chunkDependencyKeys = chunkMap[1];",
+    "var chunkModuleHashMap = chunkMap[2];",
+    "console.log({chunkBelongsToThisBuild:compilationHash === chunkCompileHash, chunkCompileHash:chunkCompileHash, chunkDependencyKeys:chunkDependencyKeys,chunkModuleHashMap:chunkModuleHashMap})",
+    "if(compilationHash !== chunkCompileHash) {",
+    "chunkDependencyKeys.forEach(function(key){",
+    "console.log('key',key,'modules',chunkModuleHashMap[key],'webpack modules',modules)",
     "//console.log(chunkMap[1][key].filter((neededModule)=>{return !modules[neededModule]}));//modules.filter((module)=>{console.log(module)})",
     // "Object.keys(chunkMap).forEach(function(key){",
     // "interleaveMap[key] = chunkMap[key]",
-    "//})",
+    "})",
+    "}",
     "};"
   ].join("\n");
 };
